도메인 설계서 참고하여 수정

+++ api-module/src/main/java/com/onde/api/application/flight/FlightController.java
@@ -105,7 +105,7 @@ public class FlightController {
 
     private void triggerCompensatingRefund(String pgTransactionId, BigDecimal amount) {
         try {
-            portOneService.cancelPayment(pgTransactionId, amount.longValue(), "로컬 DB 갱신 장애로 인한 SAGA 보상 트랜잭션 자동 취소");
+            portOneService.cancelPayment(pgTransactionId, amount, "로컬 DB 갱신 장애로 인한 SAGA 보상 트랜잭션 자동 취소");
             log.info(
                     "💰 [SAGA COMPENSATION SUCCESS] Auto-refund completed successfully via PortOne for pgTransactionId={} with amount={}",
                     pgTransactionId, amount);
