+++ api-module/src/main/java/com/onde/api/application/payment/dto/request/PaymentPrepareRequest.java
@@ -3,6 +3,8 @@ package com.onde.api.application.payment.dto.request;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * PG 결제 사전 등록(검증 준비)을 요청할 때 사용되는 DTO 클래스입니다.
  */
@@ -23,10 +25,10 @@ public class PaymentPrepareRequest {
     /**
      * 이번 결제에서 사용하겠다고 프론트엔드에서 입력한 마일리지 금액
      */
-    private Long usedMileage;
+    private Integer usedMileage;
 
     /**
      * 마일리지 적용 전, 결제되어야 할 총 상품/예약 원금액
      */
-    private Long totalAmount;
+    private BigDecimal totalAmount;
 }
